Add some colors to translated VCD file
This adds some colors to the VCD translated files by prefixing the signal names with the corresponding color.

This is specifically designed for gtkwave and requires https://github.com/gtkwave/gtkwave/pull/120 to not have those color names show up in some places. Because of that we should probably add a flag to both spade and swim to disable this, but for now I think the advantages outweigh the disadvantages since all our users use gtkwave :D